requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.6
cffi==1.12.2
Django==1.11.20
django-debug-toolbar==1.11
enum34==1.1.6
Pillow==5.4.1
pycparser==2.19
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
